requirements.txt
streamlit
google-generativeai
PyPDF2
google-generativeai>=0.5.0